`UAOOI.Configuration.DataBindings.Common` move to separate package fixes #275
