+++ Configuration/Core/InstanceNodeClassesEnum.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace UAOOI.Configuration.DataBindings.Common
+namespace UAOOI.Configuration.Core
 {
   /// <summary>
   /// Enumeration of the node classes that can be a source of process data.
