+++ Configuration/DataBindings/ConfigurationBase.cs
@@ -2,7 +2,7 @@
 
 using System;
 using System.IO;
-using UAOOI.Configuration.DataBindings.Common;
+using UAOOI.Configuration.Core;
 using UAOOI.Configuration.Networking;
 
 namespace UAOOI.Configuration.DataBindings
