+++ Configuration/DataBindings/IConfigurationEditor.cs
@@ -1,6 +1,6 @@
 ﻿
 using System;
-using UAOOI.Configuration.DataBindings.Common;
+using UAOOI.Configuration.Core;
 using UAOOI.Configuration.Networking.Serialization;
 
 namespace UAOOI.Configuration.DataBindings
