+++ Configuration/DataBindings/UANetworkingConfigurationEditor.cs
@@ -2,7 +2,7 @@
 using System;
 using System.ComponentModel.Composition;
 using System.IO;
-using UAOOI.Configuration.DataBindings.Common;
+using UAOOI.Configuration.Core;
 using UAOOI.Configuration.Networking.Serialization;
 
 namespace UAOOI.Configuration.DataBindings
