+++ Configuration/Tests/DataBindingsUnitTest/Exports/InstanceConfigurationFactory.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Linq;
-using UAOOI.Configuration.DataBindings.Common;
+using UAOOI.Configuration.Core;
 using UAOOI.Configuration.Networking.Serialization;
 
 namespace UAOOI.Configuration.DataBindings.UnitTest.Exports
